+++ AnyMark/Pattern/Patterns/AutoLink.php
@@ -6,7 +6,7 @@
 namespace AnyMark\Pattern\Patterns;
 
 use AnyMark\Pattern\Pattern;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 /**
  * @package AnyMark
@@ -44,7 +44,7 @@ class AutoLink extends Pattern
 	}
 
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {
 		$a = $this->createElement('a');
 		if (isset($match['url']))