+++ AnyMark/Pattern/Patterns/Blockquote.php
@@ -6,7 +6,7 @@
 namespace AnyMark\Pattern\Patterns;
 
 use AnyMark\Pattern\Pattern;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 /**
  * @package AnyMark
@@ -29,7 +29,7 @@ class Blockquote extends Pattern
 	}
 
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {
 		$text = preg_replace("#(^|\n)> ?#", "\${1}", $match['quote']);
 		$blockquote = $this->createElement('blockquote');