+++ AnyMark/Pattern/Patterns/CodeInline.php
@@ -6,7 +6,7 @@
 namespace AnyMark\Pattern\Patterns;
 
 use AnyMark\Pattern\Pattern;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 /**
  * @package AnyMark
@@ -24,7 +24,7 @@ class CodeInline extends Code
 	}
 
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {
 		# if code between backticks starts or ends with code between
 		# backticks: remove the spacing