+++ AnyMark/Pattern/Patterns/Image.php
@@ -7,7 +7,7 @@ namespace AnyMark\Pattern\Patterns;
 
 use AnyMark\Pattern\Pattern;
 use AnyMark\Plugins\LinkDefinitionCollector;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 /**
  * @package AnyMark
@@ -51,7 +51,7 @@ class Image extends Pattern
 	}
 
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {
 		if (isset($match['reference']))
 		{