+++ AnyMark/Pattern/Patterns/Strong.php
@@ -6,7 +6,7 @@
 namespace AnyMark\Pattern\Patterns;
 
 use AnyMark\Pattern\Pattern;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 /**
  * @package AnyMark
@@ -38,7 +38,7 @@ class Strong extends Pattern
 	}
 
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {
 		$marker = $match['marker'] . $match['marker'];
 		if (substr($match[0], 0, 2) !== $marker || substr($match[0], -2) !== $marker)