+++ EndToEndTests/Support/Patterns/FooChange.php
@@ -6,7 +6,7 @@
 namespace AnyMark\EndToEndTests\Support\Patterns;
 
 use AnyMark\Pattern\Pattern;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 /**
  * @package AnyMark
@@ -19,7 +19,7 @@ class FooChange extends \AnyMark\Pattern\Pattern
 	}
 	
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {
 		return $this->createText('bar');
 	}