+++ UnitTests/Support/DummyPattern.php
@@ -6,7 +6,7 @@
 namespace AnyMark\UnitTests\Support;
 
 use AnyMark\Pattern\Pattern;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 /**
  * @package AnyMark
@@ -17,6 +17,6 @@ class DummyPattern extends Pattern
 	{}
 	
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {}
 }