+++ UnitTests/Support/MockPatternCreatingMultiNodes.php
@@ -3,7 +3,7 @@
 namespace AnyMark\UnitTests\Support;
 
 use \AnyMark\Pattern\Pattern;
-use ElementTree\Composable;
+use ElementTree\Element;
 
 class MockPatternCreatingMultiNodes extends \AnyMark\Pattern\Pattern
 {
@@ -27,7 +27,7 @@ class MockPatternCreatingMultiNodes extends \AnyMark\Pattern\Pattern
 	}
 
 	public function handleMatch(
-		array $match, Composable $parent, Pattern $parentPattern = null
+		array $match, Element $parent = null, Pattern $parentPattern = null
 	) {
 		$el = $this->createElement($this->elementName);
 		foreach ($this->subElements as $subElement)